members/jinho/webapp/requirements.txt
fastapi
uvicorn[standard]
ultralytics>=8.3.0
Pillow>=10.0.0
numpy
opencv-python-headless
python-multipart
